backend/requirements.txt
fastapi
uvicorn
yt-dlp
sqlalchemy
pyjwt
python-multipart
